requirements.txt
# requirements.txt
# LS 하이브리드 터틀 자동매매 — 필요한 라이브러리 목록
#
# 설치 방법:
#   pip install -r requirements.txt

# LS증권 OpenAPI 래퍼 (programgarden 제공)
programgarden-finance==1.4.3

# 환경변수 .env 파일 읽기
python-dotenv

# HTTP 요청 (텔레그램 알림 발송)
requests

# 시간대 처리 (KST 한국 표준시)
pytz

# Google Sheets 연동 (체결 원장 동기화)
gspread
oauth2client
